requirements.txt
# AI Crawler - 통합 의존성 패키지
# 백엔드 + MCP 서버 통합 환경

# 웹 프레임워크
fastapi==0.115.12
uvicorn==0.32.1
websockets>=11.0.3

# HTTP 클라이언트  
httpx>=0.28.1
requests>=2.32.0

# HTML 파싱
beautifulsoup4>=4.12.3
lxml>=4.9.3

# 크롤링 엔진들
firecrawl-py>=1.9.3
crawl4ai>=0.3.74
playwright>=1.40.0
selenium>=4.27.1

# 데이터 처리
pandas>=2.2.3
numpy>=2.1.3

# AI/분석 관련 (MCP용)
openai>=1.57.2
anthropic>=0.40.0

# 유틸리티
python-dotenv>=1.0.1
aiofiles>=24.1.0
python-multipart>=0.0.12

# 개발/테스트
pytest>=8.3.4
pytest-asyncio>=0.24.0

# 추가 유틸리티
python-dateutil>=2.8.2
chardet>=5.2.0

# 비동기 처리
anyio>=4.7.0
asyncio-mqtt>=0.16.2

# 로깅 및 모니터링
loguru>=0.7.3

# MCP 서버 (OpenAI MCP 표준)
mcp>=1.1.0 